Address auto date helpers feedback.
