+++ src/legacy/core_plugins/data/public/search/aggs/agg_config.ts
@@ -397,7 +397,6 @@ export class AggConfig {
   fieldIsTimeField() {
     const indexPattern = this.getIndexPattern();
     if (!indexPattern) return false;
-    // @ts-ignore
     const timeFieldName = indexPattern.timeFieldName;
     return timeFieldName && this.fieldName() === timeFieldName;
   }
