+++ src/legacy/core_plugins/data/public/search/aggs/index.ts
@@ -27,7 +27,7 @@ export { aggTypes } from './agg_types';
 export { AggConfig } from './agg_config';
 export { AggConfigs } from './agg_configs';
 export { FieldParamType } from './param_types';
-export { getAutoIntervalFromTimeRange } from './buckets/lib/date_range';
+export { calculateAutoTimeExpression } from './buckets/lib/date_utils';
 export { MetricAggType } from './metrics/metric_agg_type';
 export { AggTypeFilters } from './filter';
 export { aggTypeFieldFilters, AggTypeFieldFilters } from './param_types/filter';
@@ -49,7 +49,7 @@ export { termsAggFilter } from './buckets/terms';
 export { isType, isStringType } from './buckets/migrate_include_exclude_format';
 export { CidrMask } from './buckets/lib/cidr_mask';
 export { convertDateRangeToString } from './buckets/date_range';
-export { toAbsoluteDates } from './buckets/lib/date_range';
+export { toAbsoluteDates } from './buckets/lib/date_utils';
 export { convertIPRangeToString } from './buckets/ip_range';
 export { aggTypeFilters, propFilter } from './filter';
 export { OptionedParamType } from './param_types/optioned';
